refactor function to rearrange data
