+++ pages/_app.js
@@ -1,10 +1,6 @@
 import GlobalStyles from "../components/GlobalStyles";
-import { useState } from "react";
-//import { words } from "../dummydata/words";
 
 function MyApp({ Component, pageProps }) {
-  // const [allWords, setAllWords] = useState(words);
-
   return (
     <>
       <GlobalStyles />
